Implement DeviceClient

+++ src/main/kotlin/com/halter/core/domain/LastLocation.kt
@@ -1,6 +1,6 @@
 package com.halter.core.domain
 
 data class LastLocation (
-  val latitude: Int,
-  val longitude: Int,
+  val latitude: Double,
+  val longitude: Double,
 )
